refactor(wallet): move nullifier-watch sync onto UserKeyChain
